feat: create endpoint for purchases operations

+++ src/main/java/com/platzi/market/persistence/entity/ComprasProductoPK.java
@@ -3,20 +3,22 @@ package com.platzi.market.persistence.entity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 
+import java.math.BigInteger;
+
 @Embeddable
 public class ComprasProductoPK {
 
     @Column(name = "id_compra")
-    private Integer idCompra;
+    private BigInteger idCompra;
 
     @Column(name = "id_producto")
     private Integer idProducto;
 
-    public Integer getIdCompra() {
+    public BigInteger getIdCompra() {
         return idCompra;
     }
 
-    public void setIdCompra(Integer idCompra) {
+    public void setIdCompra(BigInteger idCompra) {
         this.idCompra = idCompra;
     }
